# src/httpServer/swagger/api-v1.yml
openapi: 3.0.0
info:
  title: VisionBoard API
  description: API for VisionBoard - Transforming Data into Actionable insights
  version: 1.0.0
paths:
...
  /api/v1/workflow:
    get:
      summary: List all workflows
      description: Returns a list of all workflows in the system
      operationId: listWorkflows
      tags:
        - Workflows
      responses:
        '200':
          description: A list of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workflow'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/workflow/{workflowId}/execute:
    post:
      summary: Execute a workflow
      description: Executes the specified workflow with the provided input data
      operationId: executeWorkflow
      tags:
        - Workflows
      parameters:
        - name: workflowId
          in: path
          required: true
          description: The ID of the workflow to execute
          schema:
            type: string
            example: example-workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              # @TODO: Improve this schema to be more specific about the expected input
              properties:
                data:
                  oneOf:
                    - type: object
                    - type: array
                      items:
                        oneOf:
                          - type: object
                          - type: string
      responses:
        '202':
          description: Workflow executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationCompleted'
        '400':
          description: Bad request, invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Workflow is disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/project/{projectId}/gh-org:
    post:
      summary: Add a GitHub organization to a project
      description: Adds a GitHub organization to a project
      operationId: addGithubOrganization
      tags:
        - Projects
      parameters:
        - name: projectId
          in: path
          required: true
          description: The ID of the project
          schema:
            type: integer
            example: 42
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                githubOrgUrl:
                  type: string
                  pattern: '^https://github.com/[^/]+'
                  example: 'https://github.com/expressjs'
              required:
                - githubOrgUrl
      responses:
        '201':
          description: GitHub organization added successfully
          headers:
            Location:
              description: URL of the newly added GitHub organization
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GithubOrganization'
        '400':
          description: Bad request, invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: GitHub organization already exists for this project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/compliance-check/{checkId}:
    get:
      summary: Get a compliance check by ID
      description: Returns a compliance check by ID
      operationId: getComplianceCheckById
      tags:
        - Compliance Checks
      parameters:
        - name: checkId
          in: path
          required: true
          description: The ID of the check
          schema:
            type: integer
            minimum: 1
            example: 53
      responses:
        '200':
          description: A compliance check
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceCheck'
        '400':
          description: Bad request, invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Compliance check not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/compliance-check:
    get:
      summary: List all compliance checks
      description: Returns a list of all compliance checks
      operationId: listComplianceChecks
      tags:
        - Compliance Checks
      responses:
        '200':
          description: A list of compliance checks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceCheck'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/project:
    post:
      summary: Create a new project
      description: Creates a new project with the provided details
      operationId: createProject
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                  pattern: '^[a-zA-Z0-9_-]+$'
                  example: express
              required:
                - name
      responses:
        '201':
          description: Project created successfully
          headers:
            Location:
              description: URL of the newly created project
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '409':
          description: Project already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request, invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/compliance-checklist:
    get:
      summary: List all compliance checklists
      description: Returns a list of all compliance checklists
      operationId: listComplianceChecklists
      tags:
        - Compliance Checklists
      responses:
        '200':
          description: A list of compliance checklists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceChecklist'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/compliance-checklist/{checklistId}:
    get:
      summary: Get a compliance checklist by ID
      description: Returns a compliance checklist by ID
      operationId: getComplianceChecklistById
      tags:
        - Compliance Checklists
      parameters:
        - name: checklistId
          in: path
          required: true
          description: The ID of the checklist
          schema:
            type: integer
            minimum: 1
            example: 1
      responses:
        '200':
          description: A compliance checklist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceChecklist'
        '400':
          description: Bad request, invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Compliance checklist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/bulk-import:
    post:
      summary: Perform bulk import
      description: Performs a bulk import of data
      operationId: performBulkImport
      tags:
        - Bulk Import
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                id:
                  type: string
                  example: load-manual-checks
                payload:
                  # @TODO: Improve this schema to be more specific about the expected input
                  oneOf:
                    - type: object
                    - type: array
                      items:
                        oneOf:
                          - type: object
                          - type: string
      responses:
        '202':
          description: Bulk import completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationCompleted'
        '400':
          description: Bad request, invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Operation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Get operations for bulk import
      description: Returns a list of operations for bulk import
      operationId: getBulkImportOperations
      tags:
        - Bulk Import
      responses:
        '200':
          description: A list of operations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Operation'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Operation:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          example: load-manual-checks
        description:
          type: string
          example: Load manual checks from project policies
        schema:
          type: string
          description: JSON Schema for the operation
      required:
        - id
        - description
        - schema
    OperationCompleted:
      type: object
      additionalProperties: false
      properties:
        status:
          type: string
          example: completed
        result:
          type: object
          additionalProperties: false
          properties:
            message:
              type: string
              example: Workflow completed successfully
            success:
              type: boolean
              example: true
        started:
          type: string
          format: date-time
          example: "2025-02-21T18:53:00.485Z"
        finished:
          type: string
          format: date-time
          example: "2025-02-21T18:53:00.485Z"
      required:
        - status
        - result
        - started
        - finished
    Workflow:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          example: example-workflow
        description:
          type: string
          example: This is an example workflow
        isEnabled:
          type: boolean
          example: true
        isRequiredAdditionalData:
          type: boolean
          example: true
        schema:
          type: string
          description: JSO Schema for the data required for the workflow
        operations:
          type: array
          nullable: true
          items:
            type: object
            additionalProperties: false
            properties:
              id:
                type: string
                example: example-operation
              description:
                type: string
                example: This is an example operation
              schema:
                type: string
                description: JSON Schema for the operation
      required:
        - id
        - description
        - isEnabled
        - isRequiredAdditionalData
    ComplianceChecklist:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          example: 1
        author:
          type: string
          example: "OpenJS Foundation"
        title:
          type: string
          maxLength: 255
          example: "Security Compliance Guide v1.0 - Incubating"
        description:
          type: string
          example: "This checklist is for projects that are in the incubating phase and have multiple maintainers."
        code_name:
          type: string
          maxLength: 255
          example: "OpenJS-SCGv1.0-incubating"
        url:
          type: string
          example: "https://openpathfinder.com/docs/checklists/OpenJS-SCGv1.0-incubating"
        created_at:
          type: string
          format: date-time
          example: "2025-02-21T18:53:00.485Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-02-21T18:53:00.485Z"
      required:
        - id
        - author
        - title
        - description
        - code_name
        - url
        - created_at
        - updated_at
    ComplianceCheck:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          example: 53
        title:
          type: string
          maxLength: 255
          example: "Refresh dependencies with annual releases"
        description:
          type: string
          example: "Ensure dependencies are refreshed through a new release at least once annually"
        default_section_number:
          type: string
          maxLength: 255
          example: "5"
        default_section_name:
          type: string
          maxLength: 255
          example: "vulnerability management"
        code_name:
          type: string
          maxLength: 255
          example: "annualDependencyRefresh"
        default_priority_group:
          type: string
          enum: ["P0", "P1", "P2", "P3", "P4", "P5", "P6", "P7", "P8", "P9", "P10", "P11", "P12", "P13", "P14", "R0", "R1", "R2", "R3", "R4", "R5", "R6", "R7", "R8", "R9", "R10", "R11", "R12", "R13", "R14"]
          example: "P14"
        is_c_scrm:
          type: boolean
          default: false
          example: true
        implementation_status:
          type: string
          enum: ["pending", "completed"]
          default: "pending"
          example: "completed"
        # @TODO: Convert to enum when nullable values are removed
        implementation_type:
          type: string
          nullable: true
          example: "manual"
        implementation_details_reference:
          type: string
          nullable: true
          example: "https://github.com/OpenPathfinder/visionBoard/issues/112"
        details_url:
          type: string
          example: "https://openpathfinder.com/docs/checks/annualDependencyRefresh"
        created_at:
          type: string
          format: date-time
          example: "2025-02-21T18:53:00.485Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-02-21T18:53:00.485Z"
      required:
        - id
        - title
        - description
        - default_section_number
        - default_section_name
        - code_name
        - default_priority_group
        - is_c_scrm
        - implementation_status
        - details_url
        - created_at
        - updated_at
    GithubOrganization:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          example: 1
        login:
          type: string
          example: expressjs
        github_org_id:
          type: integer
          nullable: true
          example: null
        node_id:
          type: string
          nullable: true
          example: null
        url:
          type: string
          nullable: true
          example: null
        avatar_url:
          type: string
          nullable: true
          example: null
        description:
          type: string
          nullable: true
          example: null
        name:
          type: string
          nullable: true
          example: null
        company:
          type: string
          nullable: true
          example: null
        blog:
          type: string
          nullable: true
          example: null
        location:
          type: string
          nullable: true
          example: null
        twitter_username:
          type: string
          nullable: true
          example: null
        is_verified:
          type: boolean
          nullable: true
          example: null
        has_organization_projects:
          type: boolean
          nullable: true
          example: null
        has_repository_projects:
          type: boolean
          nullable: true
          example: null
        public_repos:
          type: integer
          nullable: true
          example: null
        public_gists:
          type: integer
          nullable: true
          example: null
        followers:
          type: integer
          nullable: true
          example: null
        following:
          type: integer
          nullable: true
          example: null
        html_url:
          type: string
          example: 'https://github.com/expressjs'
        total_private_repos:
          type: integer
          nullable: true
          example: null
        owned_private_repos:
          type: integer
          nullable: true
          example: null
        private_gists:
          type: integer
          nullable: true
          example: null
        disk_usage:
          type: integer
          nullable: true
          example: null
        collaborators:
          type: integer
          nullable: true
          example: null
        default_repository_permission:
          type: string
          nullable: true
          example: null
        members_can_create_repositories:
          type: boolean
          nullable: true
          example: null
        two_factor_requirement_enabled:
          type: boolean
          nullable: true
          example: null
        members_allowed_repository_creation_type:
          type: string
          nullable: true
          example: null
        members_can_create_public_repositories:
          type: boolean
          nullable: true
          example: null
        members_can_create_private_repositories:
          type: boolean
          nullable: true
          example: null
        members_can_create_internal_repositories:
          type: boolean
          nullable: true
          example: null
        members_can_create_pages:
          type: boolean
          nullable: true
          example: null
        members_can_create_public_pages:
          type: boolean
          nullable: true
          example: null
        members_can_create_private_pages:
          type: boolean
          nullable: true
          example: null
        members_can_fork_private_repositories:
          type: boolean
          nullable: true
          example: null
        web_commit_signoff_required:
          type: boolean
          nullable: true
          example: null
        deploy_keys_enabled_for_repositories:
          type: boolean
          nullable: true
          example: null
        dependency_graph_enabled_for_new_repositories:
          type: boolean
          nullable: true
          example: null
        dependabot_alerts_enabled_for_new_repositories:
          type: boolean
          nullable: true
          example: null
        dependabot_security_updates_enabled_for_new_repositories:
          type: boolean
          nullable: true
          example: null
        advanced_security_enabled_for_new_repositories:
          type: boolean
          nullable: true
          example: null
        secret_scanning_enabled_for_new_repositories:
          type: boolean
          nullable: true
          example: null
        secret_scanning_push_protection_enabled_for_new_repositories:
          type: boolean
          nullable: true
          example: null
        secret_scanning_push_protection_custom_link:
          type: string
          nullable: true
          example: null
        secret_scanning_push_protection_custom_link_enabled:
          type: boolean
          nullable: true
          example: null
        github_created_at:
          type: string
          format: date-time
          nullable: true
          example: null
        github_updated_at:
          type: string
          format: date-time
          nullable: true
          example: null
        github_archived_at:
          type: string
          format: date-time
          nullable: true
          example: null
        created_at:
          type: string
          format: date-time
          example: '2025-06-15T18:15:43.352Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-06-15T18:15:43.352Z'
        project_id:
          type: integer
          example: 1
      required:
        - id
        - login
        - html_url
        - created_at
        - updated_at
...
    ErrorObject:
      type: object
      properties:
        errorCode:
          type: string
          description: Optional machine-readable error code
          example: some.error.code
        message:
          type: string
          description: Human-readable error message
          example: Something went wrong
        path:
          type: string
          description: Optional path to the field or resource
          example: /body/field
      required:
        - message
      additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorObject'
          minItems: 1
        name:
          type: string
          description: Optional error name
          example: Error
        path:
          type: string
          description: Optional API path
          example: /api/v1/resource
        status:
          type: integer
          description: Optional HTTP status code
          example: 400
      required:
        - errors
      additionalProperties: true